backend/requirements.txt
fastapi
uvicorn
motor
pydantic<2.0
pydantic[email]
python-dotenv
stripe
python-jose
passlib[bcrypt]